fix clear command for owner

+++ cogs/general.py
@@ -226,10 +226,12 @@ class General(commands.Cog):
 
     @commands.command(name="clear", hidden=True)
     @commands.is_owner()
-    async def clear(self, ctx, num_msgs):
-        async for message in ctx.channel.history(limit=num_msgs):
-            await message.delete()
-        await ctx.send(f"{num_msgs} message(s) deleted", delete_after=10)
+    async def clear(self, ctx, msg_count: int = None):
+        if msg_count:
+            await ctx.channel.purge(limit=msg_count + 1)
+        else:
+            async for message in ctx.channel.history():
+                await message.delete()
 
 
 def setup(bot):
